% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_utr.R
\name{get_utrs}
\alias{get_utrs}
\title{get_utrs}
\usage{
get_utrs(
  gene_list,
  biomart_attribute_of_gene_list = "mgi_symbol",
  mart.object = mart
)
}
\arguments{
\item{gene_list}{\code{list} containing the genes you want the UTRs for}

\item{biomart_attribute_of_gene_list}{\code{character} of which biomart attribute your gene list based on eg. ensembl_gene_id, mgi_symbol etc.}
}
\value{
\code{list} comprised of two tibbles. One for all 5' UTRs and one for all 3' UTRs + ensembl_gene_id + mgi_symbol + ensembl_transcript_id
}
\description{
get_utrs
}
\examples{
get_utrs(gene_list, 
biomart_attribute_of_gene_list = "mgi_symbol",
mart.object = mart)
}
